Fixed bug between creating and editing data
If the model's data are $fillable then we cant edit only a one or not all piece of data at a time. If the data are $guarded then we cannot create new data.

Solution: Keep the data $fillable and at the edit view all the data are autofilled from the very start inside the textboxes. I also applied the rules for the textboxes for the update function just like in the store function.