Code fix in update function
when I would ask for the rules in the update function I couldnt update the details of a car. The reason was because in the rules the model is required and in the update function I wasn't defining the model. So when pressing the button the "model" was being set to null which can't be done and thus throwing an error.